Fix interpolation being FPS-dependent
The interpolation was producing different behaviors at different FPS values.

At 10 FPS the distance the camera moved was very short and at 300 FPS it was much larger relative to the target which was moving at the same global speed in both cases (27 m/s).

3.45233 was chosen to mimick the previous behavior running at 60 FPS.

With this change the TODO comment seems to be resolved as well.